requirements.txt
kubernetes>=29.0.0
python-dotenv>=1.0.1
requests>=2.31.0
certifi==2024.8.30
urllib3==2.2.2
idna==3.10
charset-normalizer==3.3.2
pytest==8.2.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
flask>=2.3.0
openai>=1.0.0
